+++ src/components/Profile/ProfileSection/ProfileFollow/ProfileFollow.jsx
@@ -1,15 +1,17 @@
 import useUserContext from "../../../../hooks/useUserContext";
 import { Anchor } from "./styledFollow";
 
-export default function Followers({ follow, tofrom }) {
+export default function Followers({ follow }) {
   const { user } = useUserContext();
 
   return (
     <Anchor
-      to={`/profile/${user.accountname}/${tofrom.slice(0, tofrom.length - 1)}`}
+      to={`/profile/${user.accountname}/${follow.slice(0, follow.length - 1)}`}
     >
-      <strong>{follow}</strong>
-      <small>{tofrom}</small>
+      <strong>
+        {follow === "followers" ? user.followerCount : user.followingCount}
+      </strong>
+      <small>{follow}</small>
     </Anchor>
   );
 }
